refactor: move types to protocol package
