+++ integrations/standalone/package.json
@@ -4,6 +4,7 @@
   "version": "12.0.0-next",
   "dependencies": {
     "@axonivy/variable-editor": "*",
+    "@axonivy/variable-editor-protocol": "*",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
