+++ integrations/standalone/tsconfig.json
@@ -8,5 +8,8 @@
     "outDir": "lib"
   },
   "include": ["src"],
-  "references": [{ "path": "../../packages/variable-editor/tsconfig.json" }]
+  "references": [
+    { "path": "../../packages/variable-editor/tsconfig.json" },
+    { "path": "../../packages/protocol/tsconfig.json" }
+  ]
 }
