+++ packages/protocol/src/index.ts
@@ -1,1 +1,1 @@
-// wip
+export * from './types';
