+++ packages/variable-editor/package.json
@@ -17,6 +17,7 @@
   "types": "lib/index.d.ts",
   "main": "lib/editor.js",
   "dependencies": {
+    "@axonivy/variable-editor-protocol": "~12.0.0-next",
     "@axonivy/jsonrpc": "~12.0.0-next.372",
     "@axonivy/ui-components": "~12.0.0-next.372",
     "@axonivy/ui-icons": "~12.0.0-next.372",
